Generated by Spark: When I generate a random private key on the private key page and new compressed WIF is generated on that page. That is correct behavior.
When I switch from the private key page to the public key page, the private key input on the public key page does not have the value of the compressed WIF from the private key page. That is incorrect behavior.

When I switch from the public key page to the private key page the private key input contains a different private key than previously. That is incorrect behavior.